@@ -1,62 +1,108 @@
 import { Component } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthService } from './app-service/auth-service/auth.service';
+import { timer } from 'rxjs';
+import { tap, share } from 'rxjs/operators';
+import { SubSink } from 'subsink';
+import { DataService } from 'src/app/services/data-service/data.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-
-  loggedIn : boolean;
-  username : string;
-  title = 'piBee';
-
-  constructor(private _auth : AuthService, private _route : Router) { }
-
+  subscriptions = new SubSink();
+  constructor(private _data: DataService) {}
 
   ngOnInit(): void {
-    this.isLoggedIn();    
+    timer(0, 20000)
+      .pipe(
+        tap((_) => {
+          if (!this._data.loggedInUser) {
+            this.getLoggedInUser();
+          }
+          if (this._data.loggedInUser) {
+            this.getUserProfileById(this._data.loggedInUser.id);
+          }
+        })
+      )
+      .subscribe(console.log);
+
+    timer(0, 20000)
+      .pipe(
+        share(),
+        tap((_) => this.getAllUsers())
+      )
+      .subscribe();
+
+    timer(0, 20000)
+      .pipe(tap((_) => this.getAllUsersProfiles()))
+      .subscribe();
   }
 
-  isLoggedIn() {
-    if (localStorage.getItem('isLoggedIn') == "true") {
-      this.loggedIn = true;
-      this.username = localStorage.getItem('username')
-    }
-    else if (localStorage.getItem('isLoggedIn') == "false") {
-      this.loggedIn = false;
-    }
-    return(this.loggedIn)
-  }
-
-  deactivateUser(){
-    let deactivateFormData = {
-      "user" : {
-        "username" : localStorage.getItem('username'),
-        "is_active" : false
-      }
-    }
-    this._route.navigate(['logout']);
-    this._auth.EditUserDetails(deactivateFormData).subscribe(
-      deactivateUserResponse => {
-        localStorage.setItem('isLoggedIn', 'false')
-        localStorage.setItem('username', null)
-        localStorage.setItem('password', null)
-        localStorage.setItem('user', null)
-        localStorage.setItem('Token', "")
-        console.log(deactivateUserResponse)
+  getLoggedInUser() {
+    this.subscriptions.sink = this._data.GetLoggedInUser().subscribe(
+      (loggedInUserResponse) => {
+        this._data.loggedInUser = loggedInUserResponse;
       },
 
-      deactivateUserError => {
-        console.log(deactivateUserError)
-        alert(deactivateUserError.error)
+      (loggedInUserError) => {
+        console.log(loggedInUserError);
       },
 
       () => {
-        console.log("Deactivation service called Successfully")
+        console.log('GetLoggedInUser() service called successfully');
       }
-    )
+    );
+  }
+
+  getUserProfileById(id: number) {
+    this.subscriptions.sink = this._data.GetUserProfileById(id).subscribe(
+      (getUserProfileResponse) => {
+        this._data.userProfileSubject$.next(getUserProfileResponse);
+      },
+
+      (getUserProfileError) => {
+        console.log(getUserProfileError);
+      },
+
+      () => {
+        console.log('GetUserProfileById() service called successfully');
+      }
+    );
+  }
+
+  getAllUsers() {
+    this.subscriptions.sink = this._data.GetAllUsers().subscribe(
+      (getAllUsersResponse) => {
+        // this._data.allUsersArray = getAllUsersResponse;
+        this._data.allUsersSubject$.next(getAllUsersResponse);
+        // console.log(this._data.allUsersArray);
+      },
+
+      (getAllUsersError) => {
+        console.log(getAllUsersError);
+      },
+
+      () => {
+        console.log('GetAllUsers() service called successfully');
+      }
+    );
+  }
+
+  getAllUsersProfiles() {
+    this.subscriptions.sink = this._data.GetAllUsersProfiles().subscribe(
+      (getAllUsersProfilesResponse) => {
+        this._data.allUsersProfileSubject$.next(getAllUsersProfilesResponse);
+        // this._data.allUsersProfileArray = getAllUsersProfilesResponse;
+      },
+
+      (getAllUsersProfilesError) => {
+        console.log(getAllUsersProfilesError);
+      },
+
+      () => {
+        console.log('GetAllUsersProfiles() service called successfully');
+      }
+    );
   }
 }
